renaming dk files, start working on attribute names
