+++ include/functions.h
@@ -1782,8 +1782,8 @@ void func_8008A2BC(struct _HSD_GObj* gobj);
 // ? func_8008D3A4(?);
 // ? func_8008D708(?);
 f32 func_8008D7F0(struct _Fighter*);
-// ? func_8008D8D8(?);
-// ? func_8008D8E8(?);
+f32 scaleBy154_8008D8D8(f32);  
+s32 func_8008D8E8(f32);
 void func_8008D930();
 // ? func_8008DA4C(?);
 // ? func_8008DB10(?);
