+++ src/melee/ft/chara/ftDonkey/ftdonkey6.c
@@ -27,7 +27,7 @@ void ftDonkey_800E03C0(HSD_GObj* fighterObj) {
     ftDonkeyAttributes* donkey_attr;
     func_8007D5D4(fighter);
     donkey_attr = getFtSpecialAttrs2CC(fighter);
-    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr->x0 + 7, 0, NULL, 0.0f, 1.0, 0.0f);
+    Fighter_ActionStateChange_800693AC(fighterObj, donkey_attr->action_state + 7, 0, NULL, 0.0f, 1.0, 0.0f);
     ftAnim_SetAnimRate(fighterObj, 0.0f);
     func_800CB110(fighterObj, 1, 1.0f);
 }
