+++ src/melee/ft/chara/ftDonkey/ftdonkey8.c
@@ -19,9 +19,9 @@ void ftDonkey_800E05C4(HSD_GObj* fighterObj) {
 void ftDonkey_800E05E4(HSD_GObj* fighterObj) {
     Fighter* fighter = fighterObj->user_data;
     ftDonkeyAttributes* donkey_attr = getFtSpecialAttrs2CC(fighter);
-    fighter->x2344_f32 = donkey_attr->x28;
+    fighter->x2344_f32 = donkey_attr->cargo_hold.x28_LANDING_LAG;
     donkey_attr = getFtSpecialAttrs2CC(fighter);
-    func_800D5AEC(fighterObj, donkey_attr->x0 + 8, 1, 0, fighter, 0.0f, 1.0f);
+    func_800D5AEC(fighterObj, donkey_attr->action_state + 8, 1, 0, fighter, 0.0f, 1.0f);
     ftAnim_SetAnimRate(fighterObj, 0.0f);
 }
